Add date picker to task add/edit screen
Also I added flutter_localizations so we can override native
widgets language with our own Locale. And also we can have l10n
in the future.

It was necessary to override theme in date picker builder
so the blue color exactly match. When the app have theme changing
it may be removed.