Implementation of local revision storage + inject revision into the dio
Now the api repository has field for revision getter, it can be anything
but it should be a function which returns int.

And our database has exactly what we need, so we use db repository
getter for api repository. This creates a little more coupling, but
should work pretty well.

The only thing which is left - complete sync algorithm somehow.